+++ packages/blog_ui/pubspec.yaml
@@ -10,12 +10,6 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  # Forking flutter_html because of bug that breaks on lower versions of ios
-  # https://github.com/Sub6Resources/flutter_html/issues/1314#issuecomment-1635802042
-  flutter_html: 
-    git: 
-      url: https://github.com/stefanhk31/flutter_html.git
-      ref: hotfix/incompat-regexp
   flutter_widget_from_html: ^0.15.0
   google_fonts: ^6.2.1
   html: ^0.15.4
